@@ -75,16 +81,15 @@ android {
             storePassword 'Sa4e63uCh4$robloxAndroid$'
         }
     }
-    compileSdkVersion 23
-    buildToolsVersion "23.0.2"
-    lintOptions {
-        abortOnError false
+    // Legacy build path
+    compileSdkVersion 28
+    lint {
+        // Legacy api15: allow release to proceed despite resource matrix gaps and old targetSdk
+        abortOnError = false
+        checkReleaseBuilds = false
+        disable 'MissingDefaultResource','ExtraTranslation','ExpiredTargetSdkVersion'
     }
-    enforceUniquePackageName = false;
 
-    // Future TODO: change Apache HTTP usage to URLConnection
-    // Apache HTTP deprecated since API-22, but used in ReCaptcha library
-    // To continue using the Apache HTTP APIs, declare the following:
     useLibrary 'org.apache.http.legacy'
 
     sourceSets {
