@@ -1,17 +1,26 @@
-apply plugin: 'com.android.library'
+plugins {
+    id 'com.android.library'
+}
 
 android {
-    compileSdkVersion 22
-    buildToolsVersion '22.0.1'
+    namespace 'de.hdodenhof.circleimageview'
+    compileSdkVersion 28
 
     defaultConfig {
-        minSdkVersion 9
-        targetSdkVersion 22
+        minSdkVersion 15
+        targetSdkVersion 28
+    }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+
+    lint {
+        abortOnError false
     }
 }
 
 dependencies {
-    provided 'com.android.support:support-annotations:22.1.1'
+    compileOnly 'com.android.support:support-annotations:28.0.0'
 }
-
-apply from: 'https://raw.github.com/hdodenhof/gradle-mvn-push/master/gradle-mvn-push.gradle'
