@@ -2,7 +2,7 @@ package com.roblox.hybrid.modules;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.content.LocalBroadcastManager;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 
 import com.roblox.hybrid.RBHybridCommand;
 import com.roblox.hybrid.RBHybridModule;
