@@ -2,7 +2,7 @@ package com.roblox.hybrid.modules;
 
 import android.content.Intent;
 import android.graphics.Rect;
-import android.support.v4.content.LocalBroadcastManager;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 import android.util.DisplayMetrics;
 import android.util.Log;
 import android.view.View;
